PPather: Utils: slice up into smaller classes Triangle.Data
