+++ PPather/Triangles/Data/QuadArray.cs
@@ -1,0 +1,70 @@
+﻿namespace PPather.Triangles.Data
+{
+    public class QuadArray<T>
+    {
+        private const int SIZE = 512 * 5; // Max size if SIZE*SIZE = 16M
+
+        // Jagged array
+        // pointer chasing FTL
+
+        // SIZE*(SIZE*4)
+        private T[][] arrays;
+
+        private static void getIndices(int index, out int i0, out int i1)
+        {
+            i1 = index % SIZE; index /= SIZE;
+            i0 = index % SIZE;
+        }
+
+        private void allocateAt(int i0, int i1)
+        {
+            if (arrays == null) arrays = new T[SIZE][];
+
+            T[] a1 = arrays[i0];
+            if (a1 == null) { a1 = new T[SIZE * 5]; arrays[i0] = a1; }
+        }
+
+        public void SetSize(int new_size)
+        {
+            if (arrays == null) return;
+            getIndices(new_size, out int i0, out int i1);
+            for (int i = i0 + 1; i < SIZE; i++)
+                arrays[i] = null;
+        }
+
+        public void Set(int index, T x, T y, T z, T w, T sequence)
+        {
+            getIndices(index, out int i0, out int i1);
+            allocateAt(i0, i1);
+            T[] innermost = arrays[i0];
+            i1 *= 5;
+            innermost[i1 + 0] = x;
+            innermost[i1 + 1] = y;
+            innermost[i1 + 2] = z;
+            innermost[i1 + 3] = w;
+            innermost[i1 + 4] = sequence;
+        }
+
+        public void Get(int index, out T x, out T y, out T z, out T w, out T sequence)
+        {
+            getIndices(index, out int i0, out int i1);
+
+            x = default;
+            y = default;
+            z = default;
+            w = default;
+            sequence = default;
+
+            T[] a1 = arrays[i0];
+            if (a1 == null) return;
+
+            T[] innermost = arrays[i0];
+            i1 *= 5;
+            x = innermost[i1 + 0];
+            y = innermost[i1 + 1];
+            z = innermost[i1 + 2];
+            w = innermost[i1 + 3];
+            sequence = innermost[i1 + 4];
+        }
+    }
+}
