+++ PPather/Triangles/Utils.cs
@@ -18,49 +18,17 @@
 
  */
 
-using System;
 using System.Numerics;
-using System.Collections.Generic;
-using System.Runtime.InteropServices;
-using System.Runtime.CompilerServices;
+
 using static System.MathF;
 using static System.Numerics.Vector3;
 
 namespace WowTriangles
 {
-    public class Matrix4
+    public static class Utils
     {
-        private float[,] m = new float[4, 4];
-
-        public void makeQuaternionRotate(Quaternion q)
-        {
-            m[0, 0] = 1.0f - 2.0f * q.Y * q.Y - 2.0f * q.Z * q.Z;
-            m[0, 1] = 2.0f * q.X * q.Y + 2.0f * q.W * q.Z;
-            m[0, 2] = 2.0f * q.X * q.Z - 2.0f * q.W * q.Y;
-            m[1, 0] = 2.0f * q.X * q.Y - 2.0f * q.W * q.Z;
-            m[1, 1] = 1.0f - 2.0f * q.X * q.X - 2.0f * q.Z * q.Z;
-            m[1, 2] = 2.0f * q.Y * q.Z + 2.0f * q.W * q.X;
-            m[2, 0] = 2.0f * q.X * q.Z + 2.0f * q.W * q.Y;
-            m[2, 1] = 2.0f * q.Y * q.Z - 2.0f * q.W * q.X;
-            m[2, 2] = 1.0f - 2.0f * q.X * q.X - 2.0f * q.Y * q.Y;
-            m[0, 3] = m[1, 3] = m[2, 3] = m[3, 0] = m[3, 1] = m[3, 2] = 0;
-            m[3, 3] = 1.0f;
-        }
-
-        public Vector3 mutiply(Vector3 v)
-        {
-            Vector3 o;
-            o.X = m[0, 0] * v.X + m[0, 1] * v.Y + m[0, 2] * v.Z + m[0, 3];
-            o.Y = m[1, 0] * v.X + m[1, 1] * v.Y + m[1, 2] * v.Z + m[1, 3];
-            o.Z = m[2, 0] * v.X + m[2, 1] * v.Y + m[2, 2] * v.Z + m[2, 3];
-            return o;
-        }
-    }
-
-    public unsafe class Utils
-    {
-        public static bool SegmentTriangleIntersect(Vector3 p0, Vector3 p1,
-                                                    Vector3 t0, Vector3 t1, Vector3 t2,
+        public static bool SegmentTriangleIntersect(in Vector3 p0, in Vector3 p1,
+                                                    in Vector3 t0, in Vector3 t1, in Vector3 t2,
                                                     out Vector3 I)
         {
             I = new();
@@ -104,8 +72,8 @@ namespace WowTriangles
             return true;
         }
 
-        public static float PointDistanceToSegment(Vector3 p0,
-                                           Vector3 x1, Vector3 x2)
+        public static float PointDistanceToSegment(in Vector3 p0,
+                                           in Vector3 x1, in Vector3 x2)
         {
             Vector3 L = Subtract(x2, x1); // the segment vector
             float l2 = Dot(L, L);   // square length of the segment
@@ -128,7 +96,7 @@ namespace WowTriangles
             return L3.Length();
         }
 
-        public static void GetTriangleNormal(Vector3 t0, Vector3 t1, Vector3 t2, out Vector3 normal)
+        public static void GetTriangleNormal(in Vector3 t0, in Vector3 t1, in Vector3 t2, out Vector3 normal)
         {
             Vector3 u = Subtract(t1, t0); // triangle vector 1
             Vector3 v = Subtract(t2, t0); // triangle vector 2
@@ -137,8 +105,8 @@ namespace WowTriangles
             normal = Divide(normal, l);
         }
 
-        public static float PointDistanceToTriangle(Vector3 p0,
-                                                    Vector3 t0, Vector3 t1, Vector3 t2)
+        public static float PointDistanceToTriangle(in Vector3 p0,
+                                                    in Vector3 t0, in Vector3 t1, in Vector3 t2)
         {
             Vector3 u = Subtract(t1, t0); // triangle vector 1
             Vector3 v = Subtract(t2, t0); // triangle vector 2
@@ -160,8 +128,8 @@ namespace WowTriangles
             return Min(Min(d0, d1), d2);
         }
 
-        public static bool TestBoxBoxIntersect(Vector3 box0_min, Vector3 box0_max,
-                                               Vector3 box1_min, Vector3 box1_max)
+        public static bool TestBoxBoxIntersect(in Vector3 box0_min, in Vector3 box0_max,
+                                               in Vector3 box1_min, in Vector3 box1_max)
         {
             if (box0_min.X > box1_max.X) return false;
             if (box0_min.Y > box1_max.Y) return false;
@@ -335,314 +303,14 @@ namespace WowTriangles
             return true;
         }
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static float Min3(float a, float b, float c)
+        public static float Min3(float a, float b, float c)
         {
             return Min(a, Min(b, c));
         }
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static float Max3(float a, float b, float c)
+        public static float Max3(float a, float b, float c)
         {
             return Max(a, Max(b, c));
-        }
-    }
-
-    public class SparseFloatMatrix3D<T> : SparseMatrix3D<T>
-    {
-        private const float offset = 100000f;
-        private readonly float gridSize;
-
-        public SparseFloatMatrix3D(float gridSize)
-        {
-            this.gridSize = gridSize;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private int LocalToGrid(float f)
-        {
-            return (int)((f + offset) / gridSize);
-        }
-
-        public T Get(float x, float y, float z)
-        {
-            return base.Get(LocalToGrid(x), LocalToGrid(y), LocalToGrid(z));
-        }
-
-        public bool IsSet(float x, float y, float z)
-        {
-            return base.IsSet(LocalToGrid(x), LocalToGrid(y), LocalToGrid(z));
-        }
-
-        public void Set(float x, float y, float z, T val)
-        {
-            base.Set(LocalToGrid(x), LocalToGrid(y), LocalToGrid(z), val);
-        }
-    }
-
-    public class SparseFloatMatrix2D<T> : SparseMatrix2D<T>
-    {
-        private const float offset = 100000f;
-        private readonly float gridSize;
-
-        public SparseFloatMatrix2D(float gridSize)
-            : base(0)
-        {
-            this.gridSize = gridSize;
-        }
-
-        public SparseFloatMatrix2D(float gridSize, int initialCapazity)
-            : base(initialCapazity)
-        {
-            this.gridSize = gridSize;
-        }
-
-        public float GridToLocal(int grid)
-        {
-            return (grid * gridSize) - offset;
-        }
-
-        public int LocalToGrid(float f)
-        {
-            return (int)((f + offset) / gridSize);
-        }
-
-        public (T[], int) GetAllInSquare(float min_x, float min_y,
-                                      float max_x, float max_y)
-        {
-            int sx = LocalToGrid(min_x);
-            int ex = LocalToGrid(max_x);
-
-            int sy = LocalToGrid(min_y);
-            int ey = LocalToGrid(max_y);
-
-            T[] l = new T[(int)Math.Ceiling(((ex - sx + 1) * gridSize) + ((ey - sy + 1) * gridSize))];
-            int i = 0;
-            for (int x = sx; x <= ex; x++)
-            {
-                for (int y = sy; y <= ey; y++)
-                {
-                    if (base.IsSet(x, y))
-                    {
-                        l[i++] = base.Get(x, y);
-                    }
-                }
-            }
-
-            return (l, i);
-        }
-
-        public T Get(float x, float y)
-        {
-            return base.Get(LocalToGrid(x), LocalToGrid(y));
-        }
-
-        public void Set(float x, float y, T val)
-        {
-            base.Set(LocalToGrid(x), LocalToGrid(y), val);
-        }
-    }
-
-    public class SparseMatrix2D<T>
-    {
-        private readonly Dictionary<(int x, int y), T> dict;
-
-        public SparseMatrix2D(int initialCapacity)
-        {
-            dict = new(initialCapacity);
-        }
-
-        public bool HasValue(int x, int y)
-        {
-            return dict.ContainsKey((x, y));
-        }
-
-        public T Get(int x, int y)
-        {
-            T r = default;
-            dict.TryGetValue((x, y), out r);
-            return r;
-        }
-
-        public void Set(int x, int y, T val)
-        {
-            dict[(x, y)] = val;
-        }
-
-        public bool IsSet(int x, int y)
-        {
-            return HasValue(x, y);
-        }
-
-        public void Clear(int x, int y)
-        {
-            dict.Remove((x, y));
-        }
-
-        public ICollection<T> GetAllElements()
-        {
-            return dict.Values;
-        }
-    }
-
-    public class SparseMatrix3D<T>
-    {
-        private readonly Dictionary<(int x, int y, int z), T> dict = new();
-
-        public T Get(int x, int y, int z)
-        {
-            dict.TryGetValue((x, y, z), out T r);
-            return r;
-        }
-
-        public bool IsSet(int x, int y, int z)
-        {
-            return dict.ContainsKey((x, y, z));
-        }
-
-        public void Set(int x, int y, int z, T val)
-        {
-            dict[(x, y, z)] = val;
-        }
-
-        public void Clear(int x, int y, int z)
-        {
-            dict.Remove((x, y, z));
-        }
-    }
-
-    public class TrioArray<T>
-    {
-        private const int SIZE = 1024; // Max size if SIZE*SIZE = 16M
-
-        // Jagged array
-        // pointer chasing FTL
-
-        // SIZE*(SIZE*3)
-        private T[][] arrays;
-
-        private static void getIndices(int index, out int i0, out int i1)
-        {
-            i1 = index % SIZE; index /= SIZE;
-            i0 = index % SIZE;
-        }
-
-        private void allocateAt(int i0, int i1)
-        {
-            if (arrays == null)
-                arrays = new T[SIZE][];
-
-            T[] a1 = arrays[i0];
-            if (a1 == null)
-            {
-                a1 = new T[SIZE * 3];
-                arrays[i0] = a1;
-            }
-        }
-
-        public void SetSize(int new_size)
-        {
-            if (arrays == null) return;
-            getIndices(new_size, out int i0, out _);
-            for (int i = i0 + 1; i < SIZE; i++)
-                arrays[i] = null;
-        }
-
-        public void Set(int index, T x, T y, T z)
-        {
-            getIndices(index, out int i0, out int i1);
-            allocateAt(i0, i1);
-            T[] innermost = arrays[i0];
-            i1 *= 3;
-            innermost[i1 + 0] = x;
-            innermost[i1 + 1] = y;
-            innermost[i1 + 2] = z;
-        }
-
-        public void Get(int index, out T x, out T y, out T z)
-        {
-            getIndices(index, out int i0, out int i1);
-
-            x = default;
-            y = default;
-            z = default;
-
-            T[] a1 = arrays[i0];
-            if (a1 == null) return;
-
-            T[] innermost = arrays[i0];
-            i1 *= 3;
-            x = innermost[i1 + 0];
-            y = innermost[i1 + 1];
-            z = innermost[i1 + 2];
-        }
-    }
-
-    public class QuadArray<T>
-    {
-        private const int SIZE = 512 * 5; // Max size if SIZE*SIZE = 16M
-
-        // Jagged array
-        // pointer chasing FTL
-
-        // SIZE*(SIZE*4)
-        private T[][] arrays;
-
-        private static void getIndices(int index, out int i0, out int i1)
-        {
-            i1 = index % SIZE; index /= SIZE;
-            i0 = index % SIZE;
-        }
-
-        private void allocateAt(int i0, int i1)
-        {
-            if (arrays == null) arrays = new T[SIZE][];
-
-            T[] a1 = arrays[i0];
-            if (a1 == null) { a1 = new T[SIZE * 5]; arrays[i0] = a1; }
-        }
-
-        public void SetSize(int new_size)
-        {
-            if (arrays == null) return;
-            getIndices(new_size, out int i0, out int i1);
-            for (int i = i0 + 1; i < SIZE; i++)
-                arrays[i] = null;
-        }
-
-        public void Set(int index, T x, T y, T z, T w, T sequence)
-        {
-            getIndices(index, out int i0, out int i1);
-            allocateAt(i0, i1);
-            T[] innermost = arrays[i0];
-            i1 *= 5;
-            innermost[i1 + 0] = x;
-            innermost[i1 + 1] = y;
-            innermost[i1 + 2] = z;
-            innermost[i1 + 3] = w;
-            innermost[i1 + 4] = sequence;
-        }
-
-        public void Get(int index, out T x, out T y, out T z, out T w, out T sequence)
-        {
-            getIndices(index, out int i0, out int i1);
-
-            x = default;
-            y = default;
-            z = default;
-            w = default;
-            sequence = default;
-
-            T[] a1 = arrays[i0];
-            if (a1 == null) return;
-
-            T[] innermost = arrays[i0];
-            i1 *= 5;
-            x = innermost[i1 + 0];
-            y = innermost[i1 + 1];
-            z = innermost[i1 + 2];
-            w = innermost[i1 + 3];
-            sequence = innermost[i1 + 4];
         }
     }
 }
